+++ LeaveManagement_Backend.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
 using LeaveManagement_Backend.Application.Contracts.Persistence.Interfaces;
+using LeaveManagement_Backend.Application.Exceptions;
 using LeaveManagement_Backend.Application.Features.LeaveAllocations.Requests.Commands;
 using LeaveManagement_Backend.Application.Features.LeaveTypes.Requests.Commands;
+using LeaveManagement_Backend.Domain.Entities;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -24,8 +26,14 @@ namespace LeaveManagement_Backend.Application.Features.LeaveTypes.Handlers.Comma
         }
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
-            var leaveType = await _leaveTypeRepository.Get(request.Id);
+          
+            var leaveType =  await _leaveTypeRepository.Get(request.Id);
+
+            if (leaveType == null)
+                throw new NotFoundException(nameof(leaveType), request.Id);
+
             await _leaveTypeRepository.Delete(leaveType);
+           // await _leaveTypeRepository.Save();
             return Unit.Value;
         }
